+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Implement `scroll::zoom_size_only` to disable scaling in `Scroll!` descendants and resize them instead.
 * Fix small `Text!` in a `Wrap!` not starting a new line.
 
 # 0.13.8